+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 ## 2.0.0
 * deprecated node@4.x.x
 * deprecated webpack@1.x.x
-* added `sourceMap: true|false` options 
+* added `sourceMap: true|false` options (based on postcss)